refactor: Moved modals to `components/Modals`
